# Optimized the exact filter performance