<!-- 
Are you using Knative? If you do, we would love to know!
https://github.com/knative/community/issues/new?template=ADOPTERS.yaml&title=%5BADOPTERS%5D%3A+%24%7BCOMPANY+NAME+HERE%7D
-->

Fixes #7307 

Currently, we use the attributes filter internally when running the exact filter. However, the attributes filter supports cases which the exact filter does not allow. Specifically, the current implementation of the exact filter requires that all filters have a non-empty string value for both the attribute and the desired value, while the attributes filter checks if the desired value is "", and treats that as a special "allow anything" case. By rewriting the exact filter with these more strict requirements, we see a speed up in terms of performance.

To see the current creation-time validation logic, look here: https://github.com/knative/eventing/blob/44aa15e9584bcedf12126d2f892fb185e8b41bb6/pkg/eventfilter/subscriptionsapi/exact_filter.go#L39-L44

<!-- Please include the 'why' behind your changes if no issue exists -->

## Proposed Changes

<!-- Please categorize your changes:
- :gift: Add new feature
- :bug: Fix bug
- :broom: Update or clean up current behavior
- :wastebasket: Remove feature or internal logic
-->

- Instead of using the attributes filter internally, just handle the filtering logic directly

With these changes, the performance difference is:
```
benchmark                                                                                          old ns/op     new ns/op     delta
BenchmarkExactFilter/Creation:_Pass_with_exact_match_of_id-8                                       81.4          53.4          -34.38%
BenchmarkExactFilter/Run:_Pass_with_exact_match_of_id-8                                            502           425           -15.31%
BenchmarkExactFilter/Creation:_Pass_with_exact_match_of_all_context_attributes_(except_time)-8     109           89.1          -18.44%
BenchmarkExactFilter/Run:_Pass_with_exact_match_of_all_context_attributes_(except_time)-8          1441          1326          -7.98%
BenchmarkExactFilter/Creation:_No_pass_with_exact_match_of_id_and_source-8                         84.7          61.8          -27.04%
BenchmarkExactFilter/Run:_No_pass_with_exact_match_of_id_and_source-8                              762           727           -4.50%

benchmark                                                                                          old allocs     new allocs     delta
BenchmarkExactFilter/Creation:_Pass_with_exact_match_of_id-8                                       1              1              +0.00%
BenchmarkExactFilter/Run:_Pass_with_exact_match_of_id-8                                            4              4              +0.00%
BenchmarkExactFilter/Creation:_Pass_with_exact_match_of_all_context_attributes_(except_time)-8     1              1              +0.00%
BenchmarkExactFilter/Run:_Pass_with_exact_match_of_all_context_attributes_(except_time)-8          15             15             +0.00%
BenchmarkExactFilter/Creation:_No_pass_with_exact_match_of_id_and_source-8                         1              1              +0.00%
BenchmarkExactFilter/Run:_No_pass_with_exact_match_of_id_and_source-8                              7              7              +0.00%

benchmark                                                                                          old bytes     new bytes     delta
BenchmarkExactFilter/Creation:_Pass_with_exact_match_of_id-8                                       24            8             -66.67%
BenchmarkExactFilter/Run:_Pass_with_exact_match_of_id-8                                            208           208           +0.00%
BenchmarkExactFilter/Creation:_Pass_with_exact_match_of_all_context_attributes_(except_time)-8     24            8             -66.67%
BenchmarkExactFilter/Run:_Pass_with_exact_match_of_all_context_attributes_(except_time)-8          448           448           +0.00%
BenchmarkExactFilter/Creation:_No_pass_with_exact_match_of_id_and_source-8                         24            8             -66.67%
BenchmarkExactFilter/Run:_No_pass_with_exact_match_of_id_and_source-8                              410           410           +0.00%
```

### Pre-review Checklist

<!-- If these boxes are not checked, you will be asked to complete these requirements or explain why they do not apply to your PR. -->

- [ ] **At least 80% unit test coverage**
- [ ] **E2E tests** for any new behavior
- [ ] **Docs PR** for any user-facing impact
- [ ] **Spec PR** for any new API feature
- [ ] **Conformance test** for any change to the spec

**Release Note**

<!--
:page_facing_up: If this change has user-visible impact, write a release note in the block
below. Include the string "action required" if additional action is required of
users switching to the new release, for example in case of a breaking change.

Write as if you are speaking to users, not other Knative contributors. If this
change has no user-visible impact, no release note is needed.
-->

```release-note
The exact filter now uses less memory and is faster!
```


**Docs**

<!--
:book: If this change has user-visible impact, link to an issue or PR in
https://github.com/knative/docs.
-->